refactor: splitting the transaction execution into smaller bits
